Provide a more useful message than 'Oops! Unhandled error!'.
lua_pcall just pushes the message on the stack in case of failure.
So we just have to display lua_tostring(l, -1).
The call to lua_pop is here to clean the stack and should have been done
even with the old message.

See http://www.lua.org/manual/5.1/manual.html#lua_pcall.

git-svn-id: https://microlua.googlecode.com/svn@101 d12aad66-0f56-0410-8b2c-c7101ecc2a8f