Removed calls to fatInitDefault() that were placed in some i/o functions
In newer versions of libfat, this init function resets the current 
directory to root (/), so it caused problems with paths.

Also, this is a first step to remove unnecessary custom code from Lua 
source. If we can achieve this, it'll be easier to replace Lua when 
needed

WARNING: This commit breaks the fat system because now there's no init
         at all. A later commit will fix this


git-svn-id: https://microlua.googlecode.com/svn@269 d12aad66-0f56-0410-8b2c-c7101ecc2a8f