Made some parts of main Makefile conditional based on the OS.
In previous version, you had to (un)comment parts whether your OS was
Windows or Linux. Now these parts are automatically run for the right
OS.

This is done by detecting the "OS" environment variable, which should
exist only on Windows, and contain the string "Windows".

I don't know if it's the best way to do this, but it works.
If it's ok, it could be extended to other Makefiles in the project.


git-svn-id: https://microlua.googlecode.com/svn@282 d12aad66-0f56-0410-8b2c-c7101ecc2a8f