Removed references to our DS libs from Lua source code and Makefile.
WARNING: this is the first step in moving the "ds" directory out of
Lua source tree. As a result, these Micro Lua libs are not compiled
anymore with liblua, so any call to screen, Image and the like won't 
work. Lua is now compiled as if it was only basic Lua.

Further steps will move the "ds" directory to our source tree, try to
compile the Lua additional libs from there, and enable them from our
code too.


git-svn-id: https://microlua.googlecode.com/svn@287 d12aad66-0f56-0410-8b2c-c7101ecc2a8f