Replaced os.time() from Lua with a modified version, at runtime.
Normally, the os.time() in Lua source code should work on a real DS, 
but it doesn't in emulators. It causes time to be "freezed", so timers
in Micro Lua don't work. It seems to be because it uses standard C 
time(), which uses RTC, which in turn isn't implemented in emulators.

Correct me if I'm wrong.

So I had to replace os.time() with a version of our own, based on  
timers. Actually it's the same method that was used before I tried to
clean up Lua source tree, but now it's implemented in our own code 
instead of being in Lua.


git-svn-id: https://microlua.googlecode.com/svn@291 d12aad66-0f56-0410-8b2c-c7101ecc2a8f