Left a copy of different versions of Makefiles in Lua source code.
This allows us to always have both the original versions of the Makefile
and our versions, so we could later try to get rid of ours altogether,
although I don't know if it's possible.

The compiler/linker/... binaries, directories etc. should then be 
defined outside the Makefile, I don't know how to do this.

Then all the files in Lua source tree would be untouched, and a 
replacement with a newer version would be dead easy. All you'd have to
do is recompile it once without even touching its source files.


git-svn-id: https://microlua.googlecode.com/svn@292 d12aad66-0f56-0410-8b2c-c7101ecc2a8f