Used "start /B" instead of just "start" in Windows batch files.
Sometimes the FCSR image wasn't valid when created on Windows.
I supsect it's because it's launched separately, and a "sleep 0.5" is 
used to wait for it to end, which is sometimes not enough for the image
to be fully created.
"start /B" forces the launched command to stay in the current window 
and waits for it to end before continuing.


git-svn-id: https://microlua.googlecode.com/svn@296 d12aad66-0f56-0410-8b2c-c7101ecc2a8f