Change the alpha transparency system.
The functions have changed. There still is setAlpha, but the layer index is now ALWAYS OPTIONAL, unless you want to manage them yourself. Moreover, instead of getAlpha() the function is screen.getLayer(), and it also adds screen.getAlphaLevel() to get the blending coefficient.
The usage hasn't change after this; note however that render() now reset the transparency (that is to say, it makes a screen.setAlpha(ALPHA_RESET)).

git-svn-id: https://microlua.googlecode.com/svn@409 d12aad66-0f56-0410-8b2c-c7101ecc2a8f