Add control over the LED and ability to shutdown.
You can make the LED blink as you wish.
Also add a function to put the console in sleep state, which is instantly
discarded because the screens are open (of course).
There is also a new control to detect the closing of the screens, but it
is ineffective as the console is put in sleep state before it can be
detected.

git-svn-id: https://microlua.googlecode.com/svn@490 d12aad66-0f56-0410-8b2c-c7101ecc2a8f