Add support for EFSLib. [4.6.1_a [EFS]]
The compilation process now includes the Embedded File System. It will use the folder 'efsroot', so let it empty if you want a classic MicroLua.
In any case, MicroLua will first look for its files on fat:, and then on efs:.

A new normal version name will be given once we settle the problem with DateTime on 3DS.

git-svn-id: https://microlua.googlecode.com/svn@496 d12aad66-0f56-0410-8b2c-c7101ecc2a8f