Fix System.listDirectory().
By setting the current working directory as the folder we want to list, we can bypass the stat()'s bug that messes up the informations about the item being a folder and its size.
The way the items are sorted has been changed so a modified bubble sorting is used on a table pushed on Lua's stack.