Fix the too long startup.
Now 'main()' will try to init EFS only if we know it is available, so the system doesn't search through the whole card on each boot if not needed.
This system uses a macro 'EFS' defined upon compilation (thus the compilation process has been updated). Based on this macro, the variable `System.EFS` can be checked to know whether the EFS is available or not.
Also fix calling `stat()` on a folder in the EFS which was producing an error; now it will always give 0 for a folder (just like on FAT).